code/requirements.txt
# These were the latest (working) releases when starting this project
detectron2 @ git+https://github.com/facebookresearch/detectron2@e70b9229d77aa39d85f8fa5266e6ea658e92eed3
panopticapi @ git+https://github.com/cocodataset/panopticapi@7bb4655548f98f3fedc07bf37e9040a992b054b0
wandb==0.17.2
easydict==1.11
diffusers==0.26.2
transformers==4.37.2
python-dotenv==1.0.1
accelerate==0.27.2
# segmentation-models-pytorch==0.3.3
torchseg==0.0.1a4
torcheval
torchmetrics
torchinfo==1.8.0 # This is a sctrict req, as we modify a function to support distributions.
torchtnt==0.2.2
hydra-core==1.3.2
numpy==1.26.4 # Numpy 2.0 is not yet supported by the various other packages.
hydra-submitit-launcher==1.2.0 # Used on the SLURM cluster
hydra-joblib-launcher==1.2.0 # Used on my own pc
# The following are required to ensure the ax-sweeper is working
hydra-ax-sweeper==1.2.0
gpytorch==1.8.0
pandas==1.4.0
SQLAlchemy==1.4.17
scikit-image==0.24.0
